Random updates

+++ src/System/Modules/Utils/Helpers/Helpers.php
@@ -11,7 +11,7 @@ function fixFloat($input, $precision = -1)
 {
     $input = str_replace([',', ' '], ['.', ''], $input);
 
-    return ($precision === -1 ? (double)$input : round((double)$input, $precision));
+    return ($precision === -1 ? (float)$input : round((float)$input, $precision));
 }
 
 
@@ -280,7 +280,7 @@ function validISODateTime($datetime)
             . '(\d{4})-(\d{2})-(\d{2})T' // YYYY-MM-DDT ex: 2014-01-01T
             . '(\d{2}):(\d{2}):(\d{2})'  // HH-MM-SS  ex: 17:00:00
             . '(Z|((-|\+)\d{2}:\d{2}))'  // Z or +01:00 or -01:00
-        . '$/',
+            . '$/',
         $datetime
     ) == 1;
 }
